refactor: separate bet place fields and payload
